refactor(chat): split ChatStream into src/components/chat/* (#207)
ChatStream.tsx grew to ~1460 lines mixing block renderers, long-output
viewport, diff viewer, pretty-input, stall escalation logic, and the
scroll/follow orchestrator. Split into focused modules under
src/components/chat/* with zero behavior change.

ChatStream.tsx now only contains the store wiring, the in-flight tool
tick, the AnimatePresence empty/blocks crossfade, the scroll-follow +
jump-to-latest button, and the blocks.map dispatch.

Verified by: build (tsc + webpack), full vitest run (only pre-existing
better-sqlite3 NODE_MODULE_VERSION failures remain — same on baseline),
harness-agent (9/9), harness-ui (5/5), harness-perm (9/9).
