+++ src/components/chat/EmptyState.tsx
@@ -1,0 +1,10 @@
+import { useTranslation } from '../../i18n/useTranslation';
+
+export function EmptyState() {
+  const { t } = useTranslation();
+  return (
+    <div className="h-full flex items-center justify-center px-6">
+      <div className="font-mono text-sm text-fg-tertiary">{t('chat.ready')}</div>
+    </div>
+  );
+}
